[490] Handle non-existent card directory in delete series function

+++ backend/app/core/series.py
@@ -536,8 +536,9 @@ def delete_series(
         _delete_folder(series.source_directory, log=log)
 
     # If all Cards have been deleted, delete the card directory
-    if not any(series.card_directory.iterdir()):
-        series.card_directory.rmdir()
+    if (series.card_directory.exists()
+        and not any(series.card_directory.iterdir())):
+            series.card_directory.rmdir()
 
     # Delete Series; all child objects are deleted on cascade
     log.info(f'Deleting {series}')
